Implementación de las secciones de : Presupuesto y Meta

+++ src/main/java/com/Finanzas/model/Movimiento.java
@@ -1,7 +1,7 @@
 
 package com.Finanzas.model;
 
-import java.math.BigDecimal;
+
 import java.time.LocalDate;
 
 import jakarta.persistence.Column;
@@ -38,7 +38,7 @@ public class Movimiento {
 	private String descripcion;
 	
 	@Column(name = "monto")
-	private BigDecimal monto;
+	private Double monto;
 	
 	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "id_categoria")
